+++ src/BigCommerce/Settings/Screens/Settings_Screen.php
@@ -56,22 +56,22 @@ class Settings_Screen extends Abstract_Screen {
 
 		echo '<div class="bc-settings-header__cta">';
 		printf( '<h2 class="bc-settings-header__cta-title">%s</h2>', esc_html( __( 'Manage your products.', 'bigcommerce' ) ) );
+		echo '<div class="bc-settings-header__cta-btn" data-js="bc-product-sync-button">';
+		printf( '<a href="%s" target="_blank" rel="noopener" class="bc-admin-btn bc-settings-header__manage-button">%s</a>',
+			esc_url( $this->manage_products_url() ),
+			esc_html( __( 'Manage on BigCommerce', 'bigcommerce' ) )
+		);
+		/**
+		 * Triggered after rendering the last import date in the settings header
+		 */
+		do_action( 'bigcommerce/settings/header/import_status' );
+
 		if ( $last_import_date ) {
 			printf( '<p class="bc-settings-header__cta-text">%s</p>', sprintf(
 				esc_html( __( 'Your last sync was on %s', 'bigcommerce' ) ),
 				$last_import_date
 			) );
 		}
-		echo '<div class="bc-settings-header__cta-btn" data-js="bc-product-sync-button">';
-		printf( '<a href="%s" target="_blank" rel="noopener" class="bc-admin-btn bc-settings-header__manage-button">%s</a>',
-			esc_url( $this->manage_products_url() ),
-			esc_html( __( 'Manage on BigCommerce', 'bigcommerce' ) )
-		);
-
-		/**
-		 * Triggered after rendering the last import date in the settings header
-		 */
-		do_action( 'bigcommerce/settings/header/import_status' );
 
 		echo '</div>'; // bc-settings__cta-buttons